Testing out length function on query list

+++ Sockets/WebServer/src/main/java/funHttpServer/WebServer.java
@@ -261,7 +261,7 @@ class WebServer {
             for (int i = 0; i < array.size(); i++) {
               repo = (JSONObject)array.get(i);
               own = (JSONObject)repo.get("owner");
-              id = own.get("id");
+              id = (long)own.get("id");
               owner = (String)own.get("login");
               repoNames[i] = (String)repo.get("name");
             }
